Makefile: -pthread to compile the executable
Based on the quick googling, -pthread
is the right way to enable the pthead
feature for gcc.  gcc 12 doesn't even
require the -lpthread option but seems
to support gcc 11, -pthread is needed.

And -pthread does additional flag
settings in addition to giving -lpthread,
which is the right way in gcc 11 and
prior.

autotools hides those details for us
but it's good to know those details,
especially for the learning purpose.