4.3.0 Dijkstra Banker Algorithm
is_safe method to simulate the possible
starvation with the current allocation
proposal.  If it returns true, it's safe
with the current allocation, otherwise
false, as it leads to the worker starvation,
which leads to the dead/live lock.

This is so deep that I need to go through
multiple times before the algorithm sink in.