5.3.2 I/O Selector
This set_nonblocking(true) call solves the
deadlock issue.  This is a good example of
the async bug, you can't call the blocking
function inside async tasks.

I also need to master the tracing debugging,
as it's really hard to observe the behavior
with the current tracing setup with the
FmtSubscriber type.

But at least, this wraps up section 5.3.2,
I/O Selector type.